right prompts for bash
Provide a right prompt for bash so that with the default config,
those right segments don't fall into oblivion.

Given the way they are output, there is no adjustment of the current
line for window width changes, of course, but all subsequent lines will
be adjusted to the new prompt.